code/src/requirements.txt
GitPython==3.1.44
huggingface-hub==0.29.3
jupyterlab==4.3.5
jupyterlab_pygments==0.3.0
jupyterlab_server==2.27.3
streamlit==1.43.2
pandas==2.2.3
numpy==2.0.2
torch==2.6.0
torchaudio==2.6.0
torchvision==0.21.0
sentence-transformers==3.4.1
accelerate==1.5.2
faiss-cpu==1.10.0
accelerate==1.5.2
altair==5.5.0
emoji==2.14.1
fastapi==0.115.11
filelock==3.17.0
gradio==4.44.1
python-dotenv==1.0.1
requests==2.32.3
scikit-learn==1.6.1
transformers==4.50.0